backend/requirements.txt
Django==1.9.5
django-filter==0.12.0
djangorestframework==3.3.3
djangorestframework-jsonapi==2.0.0-beta.2
inflection==0.3.1
Markdown==2.6.5
wheel==0.24.0
